Rearranging examples Refs: #7
