+++ build.gradle
@@ -120,8 +120,8 @@ tasks.register('runWelcomeUnitTests', Task) {
     description 'Run unit tests for welcome app.'
 }
 
-tasks.register('runMultiCommonUnitTests', Task) {
-    dependsOn[':examples:multi:allTests']
+tasks.register('runTimerUnitTests', Task) {
+    dependsOn[':examples:timer:allTests']
     description 'Run unit tests for welcome app.'
 }
 
@@ -142,7 +142,7 @@ task runUnitTests(
                 'runRegisterUnitTests',
                 'runLceUnitTests',
                 'runWelcomeUnitTests',
-                'runMultiCommonUnitTests'
+                'runTimerUnitTests'
         ],
         group: 'verification'
 ) {
