+++ settings.gradle
@@ -44,6 +44,6 @@ include ':examples:welcome:commonregister'
 include ':examples:welcome:commonapi'
 include ':examples:welcome:register'
 include ':examples:lce'
-include ':examples:multi:common'
+include ':examples:timer'
 include ':examples:multi:navbar'
 include ':examples:multi:parallel'
